update the session's password

+++ src/main/java/com/example/posapplicationapis/services/category/CategoryServiceImpl.java
@@ -4,13 +4,16 @@ import com.example.posapplicationapis.dto.category.CategoryDtoRequest;
 import com.example.posapplicationapis.dto.category.CategoryDtoResponse;
 import com.example.posapplicationapis.entities.Category;
 import com.example.posapplicationapis.entities.Image;
+import com.example.posapplicationapis.entities.Product;
 import com.example.posapplicationapis.repositories.CategoryRepository;
 import com.example.posapplicationapis.repositories.ImageRepository;
 import com.example.posapplicationapis.service.ImageService;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -67,12 +70,20 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Transactional
     public String deleteCategory(Long categoryId) {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found"));
         Image oldImage = category.getImage();
         category.setImage(null);
-        imageRepository.delete(oldImage);
+        if(oldImage != null){
+            imageRepository.delete(oldImage);
+        }
+        for (Product product : category.getProducts()){
+            product.setCategory(null);
+        }
+        category.setProducts(null);
+        category.setMenus(null);
 
         categoryRepository.delete(category);
         return "Category deleted successfully";
@@ -119,5 +130,17 @@ public class CategoryServiceImpl implements CategoryService {
         category.setImage(image1);
         Category updatedCategory = categoryRepository.save(category);
         return mapToDto(updatedCategory);
+    }
+
+    @Override
+    public List<CategoryDtoResponse> getAll() {
+
+        List<CategoryDtoResponse> categoryDtoResponses = new ArrayList<>();
+
+        for (Category category : categoryRepository.findAll())     {
+            categoryDtoResponses.add(mapToDto(category));
+        }
+
+        return categoryDtoResponses;
     }
 }
